requirements.txt
Flask==0.9
Flask-PyMongo==0.2.1
Flask-Login==0.1.3
